spam: consistency
