+++ pod/python/praxyk/__init__.py
@@ -1,7 +1,5 @@
 import sys
 
-sys.path.append(".")
-sys.path.append("../")
 from praxyk_python import *
 from facial_tools import *
-from get_feature import *
+from spam import *
